refac: delete lib/twitch folder and move all files in lib/
