+++ lib/twitch.js
@@ -30,7 +30,7 @@ export const getViewers = async (login) => await fetchTmi(`user/${login}/chatter
 /**
  * 
  * @param {Object} response Response of API call 
- * @return {JSON | Object} Json response || Error status and message
+ * @return {JSON || Object} Json response || Error status and message
  */
 const getContent = async (response) => {
     if (response.status >= 400) {
@@ -53,8 +53,7 @@ const fetchApi = async (endpoint) => {
 
 const fetchTmi = async (endpoint) => {
     const response = await fetch(baseUriTmi + endpoint, await getHeaders())
-    const data = getContent(response)
-    return data
+    return getContent(response)
 }
 
 /**
